Check for .project file and associated directory
virtualenv wrapper supports the mkproject command which stores the desired project's code holder in the .project file in the virtualenv folder.  I've added a few lines that check for that file, and if it contains a valid folder path with change to that folder as part of the workon command.  This is consistent with the behavior of virtualenvwrapper.